refactor(automation): split wrapper generation to separate module
