+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/updating/CreateActionUpdatePRs.kt
@@ -4,16 +4,16 @@ import io.github.typesafegithub.workflows.actionsmetadata.model.ActionCoords
 import io.github.typesafegithub.workflows.actionsmetadata.model.WrapperRequest
 import io.github.typesafegithub.workflows.actionsmetadata.model.isTopLevel
 import io.github.typesafegithub.workflows.actionsmetadata.wrappersToGenerate
-import io.github.typesafegithub.workflows.codegenerator.generation.Wrapper
-import io.github.typesafegithub.workflows.codegenerator.generation.generateWrapper
-import io.github.typesafegithub.workflows.codegenerator.metadata.fetchMetadata
-import io.github.typesafegithub.workflows.codegenerator.metadata.prettyPrint
 import io.github.typesafegithub.workflows.codegenerator.types.provideTypes
 import io.github.typesafegithub.workflows.codegenerator.versions.GithubRef
 import io.github.typesafegithub.workflows.codegenerator.versions.GithubTag
 import io.github.typesafegithub.workflows.codegenerator.versions.getGithubToken
 import io.github.typesafegithub.workflows.codegenerator.versions.httpClient
 import io.github.typesafegithub.workflows.codegenerator.versions.json
+import io.github.typesafegithub.workflows.wrappergenerator.Wrapper
+import io.github.typesafegithub.workflows.wrappergenerator.fetchMetadata
+import io.github.typesafegithub.workflows.wrappergenerator.generateWrapper
+import io.github.typesafegithub.workflows.wrappergenerator.prettyPrint
 import io.ktor.client.request.bearerAuth
 import io.ktor.client.request.get
 import io.ktor.client.statement.bodyAsText
