+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/versions/SuggestVersions.kt
@@ -4,7 +4,7 @@ import io.github.typesafegithub.workflows.actionsmetadata.model.ActionCoords
 import io.github.typesafegithub.workflows.actionsmetadata.model.Version
 import io.github.typesafegithub.workflows.actionsmetadata.model.isTopLevel
 import io.github.typesafegithub.workflows.actionsmetadata.wrappersToGenerate
-import io.github.typesafegithub.workflows.codegenerator.metadata.prettyPrint
+import io.github.typesafegithub.workflows.wrappergenerator.prettyPrint
 import java.io.File
 
 /**
