+++ automation/wrapper-generator/src/main/kotlin/io/github/typesafegithub/workflows/wrappergenerator/Generation.kt
@@ -1,4 +1,4 @@
-package io.github.typesafegithub.workflows.codegenerator.generation
+package io.github.typesafegithub.workflows.wrappergenerator
 
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
@@ -17,12 +17,8 @@ import com.squareup.kotlinpoet.asTypeName
 import io.github.typesafegithub.workflows.actionsmetadata.model.ActionCoords
 import io.github.typesafegithub.workflows.actionsmetadata.model.StringTyping
 import io.github.typesafegithub.workflows.actionsmetadata.model.Typing
-import io.github.typesafegithub.workflows.codegenerator.generation.Properties.CUSTOM_INPUTS
-import io.github.typesafegithub.workflows.codegenerator.generation.Properties.CUSTOM_VERSION
-import io.github.typesafegithub.workflows.codegenerator.metadata.Input
-import io.github.typesafegithub.workflows.codegenerator.metadata.Metadata
-import io.github.typesafegithub.workflows.codegenerator.metadata.fetchMetadata
-import io.github.typesafegithub.workflows.codegenerator.metadata.prettyPrint
+import io.github.typesafegithub.workflows.wrappergenerator.Properties.CUSTOM_INPUTS
+import io.github.typesafegithub.workflows.wrappergenerator.Properties.CUSTOM_VERSION
 
 data class Wrapper(
     val kotlinCode: String,
