+++ automation/wrapper-generator/src/main/kotlin/io/github/typesafegithub/workflows/wrappergenerator/TypingGeneration.kt
@@ -1,4 +1,4 @@
-package io.github.typesafegithub.workflows.codegenerator.generation
+package io.github.typesafegithub.workflows.wrappergenerator
 
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FunSpec
