+++ automation/wrapper-generator/src/test/kotlin/io/github/typesafegithub/workflows/wrappergenerator/ClassNamingTest.kt
@@ -1,4 +1,4 @@
-package io.github.typesafegithub.workflows.codegenerator.generation
+package io.github.typesafegithub.workflows.wrappergenerator
 
 import io.github.typesafegithub.workflows.actionsmetadata.model.ActionCoords
 import io.kotest.core.spec.style.FunSpec
