+++ automation/wrapper-generator/src/test/kotlin/io/github/typesafegithub/workflows/wrappergenerator/wrappersfromunittests/ActionWithInputsSharingTypeV3.kt
