+++ automation/wrapper-generator/src/test/kotlin/io/github/typesafegithub/workflows/wrappergenerator/wrappersfromunittests/ActionWithSomeOptionalInputsV3.kt
