+++ automation/wrapper-generator/src/test/kotlin/io/github/typesafegithub/workflows/wrappergenerator/wrappersfromunittests/SimpleActionWithListsV3.kt
