+++ automation/wrapper-generator/src/test/kotlin/io/github/typesafegithub/workflows/wrappergenerator/wrappersfromunittests/SimpleActionWithRequiredStringInputsV3Test.kt
@@ -1,8 +1,8 @@
-package io.github.typesafegithub.workflows.codegenerator.generation.wrappersfromunittests
+package io.github.typesafegithub.workflows.wrappergenerator.wrappersfromunittests
 
+import io.github.typesafegithub.workflows.actions.johnsmith.SimpleActionWithRequiredStringInputsV3
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
-import io.github.typesafegithub.workflows.actions.johnsmith.SimpleActionWithRequiredStringInputsV3
 
 class SimpleActionWithRequiredStringInputsV3Test : DescribeSpec({
     it("renders with defaults") {
