+++ automation/wrapper-generator/src/test/resources/types/action-types.yml
