+++ automation/wrapper-generator/src/test/resources/types/action.yml
